Competition Ready
This version of this robot project is tuned and ready to compete in the FIRST Robotics Competition. This code was tested at the Ellensburg, WA, at the Central Washington University District Event. This code implements computer vision to place a single gear during autonomous from multiple starting positions. An incidence angle is sent from an onboard raspberry pi via UDP.